requirements.txt
lxml==6.0.2
python-docx==1.2.0
typing_extensions==4.15.0
inotify_simple
